address generator + tests
