+++ src/test/java/com/cryptowallet/WalletControllerTest.java
@@ -1,7 +1,6 @@
 package com.cryptowallet;
 
 import com.cryptowallet.controller.WalletController;
-import com.cryptowallet.entities.User;
 import com.cryptowallet.services.AddressService;
 import com.cryptowallet.services.CurrencyService;
 import com.cryptowallet.services.UserService;
@@ -13,9 +12,6 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-
 
 @RunWith(SpringRunner.class)
 @WebMvcTest(controllers = {WalletController.class})
